refactor: move string literals into a dedicated module
This functionality is relatively self-contained meta-progamming, moving
it into its own module helps to clearly delineate the metaprogramming
from the core functionality of the package.
