+++ Project.toml
@@ -8,6 +8,7 @@ Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
 CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
 CodecZlib = "0.7"
@@ -18,4 +19,5 @@ DocStringExtensions = "0.8, 0.9"
 # The wider compatibility range for `HTTP` ensures compatibility with
 # up-to-date versions of `Pluto`, etc.
 HTTP = "0.8, 0.9, 1"
+Reexport = "1.2.2"
 julia = "1.6"
